+++ app/envConfig.ts
@@ -1,4 +1,4 @@
 import { loadEnvConfig } from '@next/env'
- 
+
 const projectDir = process.cwd()
 loadEnvConfig(projectDir)
